refactor(netxlite): split http code into multiple files
This diff splits http.go and http_test.go into multiple files
to create the logical and mental space required to implement
new construction routines suitable for the engine package.

The history of this diff is simple. I started working on forking
http.go for the ./internal/enginenetx package. Then, I realized
it would have been possible to have those changes directly inside
netxlite, if only I did refactor the http implementation to have
a bit more clarity and file-based modularity. And so I did.

Part of https://github.com/ooni/probe/issues/2531
